fix(ex-ray): review round 4 - validate remoteAddr/localAddr, reject empty ech, split option-validation tests into options_test.go
